requirements.txt
# Just lint deps for pre-commit
#   for use with dependabot
#   .pre-commit-config.yaml need to be updated manually
flake8==6.1.0
dlint~=0.14.1
flake8-async~=22.11.14
flake8-bandit~=4.1.1
flake8-bugbear~=23.7.10
flake8-comprehensions~=3.14.0
flake8-eradicate~=1.5.0
flake8-mock==0.4
flake8-noqa~=1.3.2
flake8-pie~=0.16.0
flake8-pytest-style~=1.7.2
flake8-self~=0.2.2
flake8-simplify~=0.20.0
flake8-type-checking~=2.4.1
flake8-typing-imports~=1.14.0
flake8-unused-arguments~=0.0.13
flake8-warnings~=0.4.0
pep8-naming~=0.13.3
